New pypicasa app! Upload pictures using picasa. Not working. Currently on development. New rename.py! Rename ton of files! Just rename.py "something" "another" and will replace all "something" word on files for "another"
Megadownload fix. Using wget now. You can resume downloads!

+++ trunk/python/rename.py
@@ -9,13 +9,22 @@
 
 import dircache,sys,os
 
+dir = dircache.listdir(".")
+
 if len(sys.argv) < 2:
   print('Usage: rename.py "what you want change" "the real change')
+  print('Using rename.py "" "change" you will add changeFILE1, changeFILE2...')
   sys.exit(0)
-
-dir = dircache.listdir(".")
-for x in dir:
-  loc = x.find(sys.argv[1])
-  if loc != -1:
-    os.rename(x,x.replace(sys.argv[1],sys.argv[2]))
+if sys.argv[1] == "":
+  y = raw_input('Are you sure... you will append ' + sys.argv[2] + ' on all files. Y/n:').capitalize()
+  if y == "N":
+    sys.exit(0)
+  else:
+    for x in dir:
+      os.rename(x,sys.argv[2] + x)
+else:
+  for x in dir:
+    loc = x.find(sys.argv[1])
+    if loc != -1:
+      os.rename(x,x.replace(sys.argv[1],sys.argv[2]))
